+++ api-tests/DateTime.test.js
@@ -1,11 +1,9 @@
 const { multiAdapterRunners, setupServer, graphqlRequest } = require('@keystonejs/test-utils');
 const { Text, DateTime } = require('@keystonejs/fields');
-const cuid = require('cuid');
 
 function setupKeystone(adapterName) {
   return setupServer({
     adapterName,
-    name: `ks5-testdb-${cuid()}`,
     createLists: keystone => {
       keystone.createList('Post', {
         fields: {