+++ packages/app-admin-ui/client/components/PageError.js
@@ -1,7 +1,7 @@
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
 
-import { StopIcon } from '@arch-ui/icons';
+import { StopIcon } from '@primer/octicons-react';
 import { colors } from '@arch-ui/theme';
 import { Container } from '@arch-ui/layout';