+++ packages/arch/packages/pagination/src/Pagination.js
@@ -2,7 +2,7 @@
 
 import { jsx } from '@emotion/core';
 import { useEffect, useState } from 'react';
-import { ChevronLeftIcon, ChevronRightIcon, ListOrderedIcon } from '@arch-ui/icons';
+import { ChevronLeftIcon, ChevronRightIcon, KebabHorizontalIcon } from '@primer/octicons-react';
 import { FlexGroup } from '@arch-ui/layout';
 import { LoadingSpinner } from '@arch-ui/loading';
 
@@ -46,91 +46,84 @@ const Pagination = ({
   onChange,
   ...props
 }) => {
-  const [allPagesVisible, setAllPagesVisible] = useState(false);
+  if (total <= pageSize) return null;
 
-  const toggleAllPages = () => {
-    setAllPagesVisible(wereAllPagesVisible => !wereAllPagesVisible);
+  const pages = [];
+  const totalPages = Math.ceil(total / pageSize);
+
+  let minPage = 1;
+  let maxPage = totalPages;
+
+  if (limit && limit < totalPages) {
+    const rightLimit = Math.floor(limit / 2);
+    const leftLimit = rightLimit + (limit % 2) - 1;
+    minPage = currentPage - leftLimit;
+    maxPage = currentPage + rightLimit;
+
+    if (minPage < 1) {
+      maxPage = limit;
+      minPage = 1;
+    }
+
+    if (maxPage > totalPages) {
+      minPage = totalPages - limit + 1;
+      maxPage = totalPages;
+    }
+  }
+
+  const handleChange = page => {
+    if (onChange) {
+      onChange(page, {
+        pageSize,
+        total,
+        minPage,
+        maxPage,
+      });
+    }
   };
 
-  const renderPages = () => {
-    if (total <= pageSize) return [];
+  // go to first
+  if (minPage > 1) {
+    pages.push(
+      <Page aria-label={ariaPageLabel(1)} key="page_start" onClick={handleChange} value={1}>
+        <KebabHorizontalIcon />
+      </Page>
+    );
+  }
 
-    let pages = [];
-    let totalPages = Math.ceil(total / pageSize);
-    let minPage = 1;
-    let maxPage = totalPages;
-    const moreCharacter = <span>&hellip;</span>;
+  // loop over range
+  for (let page = minPage; page <= maxPage; page++) {
+    const isSelected = page === currentPage;
+    pages.push(
+      <Page
+        aria-label={ariaPageLabel(page)}
+        aria-current={isSelected ? 'page' : null}
+        key={`page_${page}`}
+        isSelected={isSelected}
+        onClick={handleChange}
+        value={page}
+      >
+        <PageChildren isLoading={isLoading} page={page} isSelected={isSelected} />
+      </Page>
+    );
+  }
 
-    if (limit && limit < totalPages) {
-      let rightLimit = Math.floor(limit / 2);
-      let leftLimit = rightLimit + (limit % 2) - 1;
-      minPage = currentPage - leftLimit;
-      maxPage = currentPage + rightLimit;
+  // go to last
+  if (maxPage < totalPages) {
+    pages.push(
+      <Page
+        aria-label={ariaPageLabel(totalPages)}
+        key="page_end"
+        onClick={handleChange}
+        value={totalPages}
+      >
+        <KebabHorizontalIcon />
+      </Page>
+    );
+  }
 
-      if (minPage < 1) {
-        maxPage = limit;
-        minPage = 1;
-      }
-      if (maxPage > totalPages) {
-        minPage = totalPages - limit + 1;
-        maxPage = totalPages;
-      }
-    }
-
-    const handleChange = page => {
-      if (onChange) {
-        setAllPagesVisible(false);
-        onChange(page, {
-          pageSize,
-          total,
-          minPage,
-          maxPage,
-        });
-      }
-    };
-
-    // go to first
-    if (minPage > 1) {
-      pages.push(
-        <Page aria-label={ariaPageLabel(1)} key="page_start" onClick={handleChange} value={1}>
-          {moreCharacter}
-        </Page>
-      );
-    }
-
-    // loop over range
-    for (let page = minPage; page <= maxPage; page++) {
-      const isSelected = page === currentPage;
-      pages.push(
-        <Page
-          aria-label={ariaPageLabel(page)}
-          aria-current={isSelected ? 'page' : null}
-          key={`page_${page}`}
-          isSelected={isSelected}
-          onClick={handleChange}
-          value={page}
-        >
-          <PageChildren isLoading={isLoading} page={page} isSelected={isSelected} />
-        </Page>
-      );
-    }
-
-    // go to last
-    if (maxPage < totalPages) {
-      pages.push(
-        <Page
-          aria-label={ariaPageLabel(totalPages)}
-          key="page_end"
-          onClick={handleChange}
-          value={totalPages}
-        >
-          {moreCharacter}
-        </Page>
-      );
-    }
-
-    // return pages;
-    return [
+  return (
+    <FlexGroup as="nav" align="center" aria-label="Pagination" isInline {...props}>
       <Page
         aria-label="Go to previous page"
         key="page_prev"
@@ -139,20 +132,8 @@ const Pagination = ({
         isDisabled={currentPage === 1}
       >
         <ChevronLeftIcon />
-      </Page>,
-      allPagesVisible ? (
-        pages
-      ) : (
-        <Page
-          aria-label="Click to show all pages"
-          key="page_dot"
-          onClick={toggleAllPages}
-          id="ks-pagination-show-pages"
-          value={1} // needs value for flow...
-        >
-          <ListOrderedIcon />
-        </Page>
-      ),
+      </Page>
+      {pages}
       <Page
         aria-label="Go to next page"
         key="page_next"
@@ -161,13 +142,7 @@ const Pagination = ({
         isDisabled={currentPage === totalPages}
       >
         <ChevronRightIcon />
-      </Page>,
-    ];
-  };
-
-  return (
-    <FlexGroup as="nav" align="center" aria-label="Pagination" isContiguous isInline {...props}>
-      {renderPages()}
+      </Page>
     </FlexGroup>
   );
 };