+++ packages/fields/src/types/CalendarDay/Implementation.js
@@ -67,17 +67,17 @@ export class CalendarDay extends Implementation {
     const parsedValue = parseISO(resolvedData[this.path]);
 
     if (!(initialValue.length === 10 && isValid(parsedValue))) {
-      addFieldValidationError('Invalid CalendarDay value', { value: resolvedData[this.path] });
+      addFieldValidationError('Invalid CalendarDay value.', { value: resolvedData[this.path] });
     }
     if (parsedValue) {
       if (parseISO(this._dateFrom) && compareAsc(parseISO(this._dateFrom), parsedValue) === 1) {
-        addFieldValidationError('Value is before earliest allowed date.', {
+        addFieldValidationError(`Value is before earliest allowed date: ${this._dateFromString}.`, {
           value: resolvedData[this.path],
           dateFrom: this._dateFromString,
         });
       }
       if (parseISO(this._dateTo) && compareDesc(parseISO(this._dateTo), parsedValue) === 1) {
-        addFieldValidationError('Value is after latest allowed date.', {
+        addFieldValidationError(`Value is after latest allowed date: ${this._dateToString}.`, {
           value: resolvedData[this.path],
           dateTo: this._dateToString,
         });