+++ test-projects/basic/index.js
@@ -43,8 +43,7 @@ const { formatISO } = require('date-fns');
 const { MongooseAdapter } = require('@keystonejs/adapter-mongoose');
 
 const keystone = new Keystone({
-  name: 'Cypress Test Project Basic',
-  adapter: new MongooseAdapter(),
+  adapter: new MongooseAdapter({ mongoUri: 'mongodb://localhost/cypress-test-project' }),
   cookieSecret: 'qwerty',
 });
 
@@ -275,6 +274,6 @@ module.exports = {
   apps: [
     new GraphQLApp(),
     new StaticApp({ path: staticRoute, src: staticPath }),
-    new AdminUIApp({ enableDefaultRoute: true }),
+    new AdminUIApp({ name: 'Cypress Test Project Basic', enableDefaultRoute: true }),
   ],
 };